requirements.txt
PyQt5
pynput
pywin32
pywinauto
Pillow
groq
send2trash
pystray
python-dotenv
pyperclip
psutil
opencv-python
pyzbar
pycaw
rembg
comtypes
screen-brightness-control
pytesseract
